% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_enrichment.R
\name{GO_enrichment}
\alias{GO_enrichment}
\title{GO_enrichment}
\usage{
GO_enrichment(gene_vector, sub_ontology, term2name, pval, qval)
}
\description{
A small function that performs GO enrichment.
Built upon the clusterProfiler enricher function
}
\details{
@param gene_vector character vector of genes to perform enrichment on.
 @param sub_ontology GO sub-onology to perform enrichment on. This is stored in the object list returned by prep_data() included in this package.
 @param term2name the dataframe linking gene names to GO terms. Required by enricher from clusterProfiler. This is stored in the object list returned by prep_data() included in this package
 @param pval ajusted p-value to threshold enriched terms on, same as pvalueCutoff from clusterProfiler
 @param qval FDR to threshold enriched terms on, same as qvalueCutoff from clusterProfiler
 @return clusterProfiler enrichment result

 @examples
 GO_enrichment("dge_genes", prep_data.list$GO_BP_TERMS, prep_data.list$GOTERM2LOC, 0.05, 0.2)

 @export
}
