% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_enrichment.R
\name{KEGG_enrichment}
\alias{KEGG_enrichment}
\title{KEGG_enrichment}
\usage{
KEGG_enrichment(gene_vector, KEGGTERM2LOC, pval, qval)
}
\description{
A small function that performs GO enrichment.
Built upon the clusterProfiler enrichKEGG function
}
\details{
@param gene_vector character vector of genes to perform enrichment on.
 @param KEGGTERM2LOC Mapping of KEGG terms to gene ID. This is stored in the object list returned by prep_data() included in this package.
 @param pval ajusted p-value to threshold enriched terms on, same as pvalueCutoff from clusterProfiler
 @param qval FDR to threshold enriched terms on, same as qvalueCutoff from clusterProfiler
 @return clusterProfiler enrichment result

 @examples
 KEGG_enrichment("dge_genes", prep_data.list$KEGGTERm2LOC,0.05, 0.2)

 @export
}
