% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{lift_annotaitons}
\alias{lift_annotaitons}
\title{Lift Annotations from Eggnog & Create GO/KEGG Tables}
\usage{
lift_annotaitons(eggnog_file, gff_file)
}
\description{
A function that replaces protein IDs from Schistocerca with LOCID from RefSeq Annotation.
It also creates GO/KEGG annotation tables for use in pathway enrichment.
}
\details{
@param eggnog_file .annotaitons file output from the EggNog Mapper tool.
 @param gff_file .gff file downloaded from NCBI FTP server
 @return a named list of objects needed for downstream pathway enrichment

 @examples
 lift_annotations("schisctocerca_greg.annotations", "GCF_023897955.1_iqSchGreg1.2_genomic.gff")

 @export
}
